renaming strategy into erc721-with-balance
