+++ src/strategies/erc721-with-balance/schema.json
