+++ src/strategies/index.ts
@@ -291,7 +291,7 @@ import * as meebitsDaoDelegation from './meebitsdao-delegation';
 import * as starcatchersTopWindow from './starcatchers-top-window';
 import * as gno from './gno';
 import * as umaVoting from './uma-voting';
-import * as erc721OneVotePerAccountBalance from './erc721-one-vote-per-account-balance';
+import * as erc721WithBalance from './erc721-with-balance';
 
 const strategies = {
   'landdao-token-tiers': landDaoTiers,
@@ -345,7 +345,7 @@ const strategies = {
   'erc721-with-tokenid-weighted': erc721WithTokenIdWeighted,
   'erc721-with-metadata': erc721WithMetadata,
   'erc721-multi-registry': erc721MultiRegistry,
-  'erc721-one-vote-per-account-balance': erc721OneVotePerAccountBalance,
+  'erc721-with-balance': erc721WithBalance,
   'erc1155-balance-of': erc1155BalanceOf,
   'erc1155-balance-of-cv': erc1155BalanceOfCv,
   multichain,
